Seperate Cmdr Calculations from MmrService
